pyproject.toml
[tool.poetry]
name = "mtu voting system"
version = "0.1.0"
description = ""
authors = ["dakohhh <wiizzydreadmill@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
fastapi = "^0.109.2"
mongoengine = "^0.27.0"
cloudinary = "^1.38.0"
uvicorn = "^0.27.1"
certifi = "^2024.2.2"
bcrypt = "^4.1.2"
python-dotenv = "^1.0.1"
PyJWT = "^2.8.0"
email-validator = "^2.1.0.post1"
fastapi-mail = "^1.4.1"
beanie = "^1.25.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
